+++ NEWS.md
@@ -1,5 +1,7 @@
 # quarto (development version)
 
+- Add `quarto_bin_sitrep()` to check possible difference in Quarto binary used by this package, and the one used by RStudio IDE (thanks, @jthomasmock, #12).
+
 - Add `quarto_create_project()` function that calls `quarto create project <type> <name>` (thanks, @maelle, #87).
 
 - Add `is_using_quarto()` to check if a directory requires using Quarto (i.e. it has a `_quarto.yml` or at least one `*.qmd` file) (thanks, @hadley, #103).